In-Memory checkpointing kind of working
