+++ src/include/duckdb/storage/checkpoint_manager.hpp
@@ -133,30 +133,4 @@ private:
 	unordered_map<block_id_t, idx_t> verify_block_usage_count;
 };
 
-class InMemoryCheckpointer final : public CheckpointWriter {
-public:
-	InMemoryCheckpointer(QueryContext context, AttachedDatabase &db, BlockManager &block_manager);
-
-	void CreateCheckpoint() override;
-
-	MetadataWriter &GetMetadataWriter() override;
-	MetadataManager &GetMetadataManager() override;
-	unique_ptr<TableDataWriter> GetTableDataWriter(TableCatalogEntry &table) override;
-	optional_ptr<ClientContext> GetClientContext() const {
-		return context;
-	}
-	PartialBlockManager &GetPartialBlockManager() {
-		return partial_block_manager;
-	}
-
-public:
-	void WriteTable(TableCatalogEntry &table, Serializer &serializer) override;
-
-private:
-	optional_ptr<ClientContext> context;
-	//! Because this is single-file storage, we can share partial blocks across
-	//! an entire checkpoint.
-	PartialBlockManager partial_block_manager;
-};
-
 } // namespace duckdb
