+++ src/include/duckdb/storage/table/column_checkpoint_state.hpp
@@ -87,7 +87,7 @@ public:
 	bool IsFlushed();
 	void Flush(QueryContext context, const idx_t free_space_left) override;
 	void Merge(PartialBlock &other, idx_t offset, idx_t other_size) override;
-	void AddSegmentToTail(ColumnData &data, ColumnSegment &segment, uint32_t offset_in_block);
+	void AddSegmentToTail(ColumnData &data, ColumnSegment &segment, uint32_t offset_in_block) override;
 	void Clear() override;
 };
