+++ src/storage/checkpoint/row_group_writer.cpp
@@ -30,25 +30,4 @@ optional_ptr<MetadataManager> SingleFileRowGroupWriter::GetMetadataManager() {
 	return table_data_writer.GetManager();
 }
 
-InMemoryRowGroupWriter::InMemoryRowGroupWriter(TableCatalogEntry &table, PartialBlockManager &partial_block_manager,
-                                               TableDataWriter &writer)
-    : RowGroupWriter(table, partial_block_manager), writer(writer) {
-}
-
-CheckpointType InMemoryRowGroupWriter::GetCheckpointType() const {
-	return CheckpointType::FULL_CHECKPOINT;
-}
-
-WriteStream &InMemoryRowGroupWriter::GetPayloadWriter() {
-	return metadata_writer;
-}
-
-MetaBlockPointer InMemoryRowGroupWriter::GetMetaBlockPointer() {
-	return MetaBlockPointer();
-}
-
-optional_ptr<MetadataManager> InMemoryRowGroupWriter::GetMetadataManager() {
-	return nullptr;
-}
-
 } // namespace duckdb
