+++ src/storage/checkpoint_manager.cpp
@@ -601,57 +601,4 @@ void CheckpointReader::ReadTableData(CatalogTransaction transaction, Deserialize
 	bound_info.data->total_rows = total_rows;
 }
 
-//===--------------------------------------------------------------------===//
-// In-Memory Checkpoint Writer
-//===--------------------------------------------------------------------===//
-InMemoryCheckpointer::InMemoryCheckpointer(QueryContext context, AttachedDatabase &db, BlockManager &block_manager)
-    : CheckpointWriter(db), context(context.GetClientContext()),
-      partial_block_manager(context, block_manager, PartialBlockType::APPEND_TO_TABLE) {
-}
-
-void InMemoryCheckpointer::CreateCheckpoint() {
-	vector<reference<SchemaCatalogEntry>> schemas;
-	// we scan the set of committed schemas
-	auto &catalog = Catalog::GetCatalog(db).Cast<DuckCatalog>();
-	catalog.ScanSchemas([&](SchemaCatalogEntry &entry) { schemas.push_back(entry); });
-
-	vector<reference<TableCatalogEntry>> tables;
-	for (const auto &schema_ref : schemas) {
-		auto &schema = schema_ref.get();
-		schema.Scan(CatalogType::TABLE_ENTRY, [&](CatalogEntry &entry) {
-			if (entry.type == CatalogType::TABLE_ENTRY) {
-				tables.push_back(entry.Cast<TableCatalogEntry>());
-			}
-		});
-	}
-
-	for (auto &table : tables) {
-		MemoryStream write_stream;
-		BinarySerializer serializer(write_stream);
-
-		WriteTable(table, serializer);
-	}
-}
-
-MetadataWriter &InMemoryCheckpointer::GetMetadataWriter() {
-	throw InternalException("Unsupported method GetMetadataWriter for InMemoryCheckpointer");
-}
-MetadataManager &InMemoryCheckpointer::GetMetadataManager() {
-	throw InternalException("Unsupported method GetMetadataManager for InMemoryCheckpointer");
-}
-unique_ptr<TableDataWriter> InMemoryCheckpointer::GetTableDataWriter(TableCatalogEntry &table) {
-	throw InternalException("Unsupported method GetTableDataWriter for InMemoryCheckpointer");
-}
-
-void InMemoryCheckpointer::WriteTable(TableCatalogEntry &table, Serializer &serializer) {
-	InMemoryTableDataWriter data_writer(*this, table);
-
-	// Write the table data
-	auto table_lock = table.GetStorage().GetCheckpointLock();
-	table.GetStorage().Checkpoint(data_writer, serializer);
-	// flush any partial blocks BEFORE releasing the table lock
-	// flushing partial blocks updates where data lives and is not thread-safe
-	partial_block_manager.FlushPartialBlocks();
-}
-
 } // namespace duckdb
