+++ src/storage/partial_block_manager.cpp
@@ -15,6 +15,10 @@ void PartialBlock::AddUninitializedRegion(idx_t start, idx_t end) {
 	uninitialized_regions.push_back({start, end});
 }
 
+void PartialBlock::AddSegmentToTail(ColumnData &data, ColumnSegment &segment, uint32_t offset_in_block) {
+	throw InternalException("PartialBlock::AddSegmentToTail not supported for this block type");
+}
+
 void PartialBlock::FlushInternal(const idx_t free_space_left) {
 
 	// ensure that we do not leak any data
@@ -103,6 +107,14 @@ bool PartialBlockManager::GetPartialBlock(idx_t segment_size, unique_ptr<Partial
 	D_ASSERT(partial_block->state.offset > 0);
 	D_ASSERT(ValueIsAligned(partial_block->state.offset));
 	return true;
+}
+
+unique_ptr<PartialBlock> PartialBlockManager::CreatePartialBlock(ColumnData &column_data, ColumnSegment &segment,
+                                                                 PartialBlockState state, BlockManager &block_manager) {
+	if (partial_block_type == PartialBlockType::IN_MEMORY_CHECKPOINT) {
+		return make_uniq<InMemoryPartialBlock>(column_data, segment, state, block_manager);
+	}
+	return make_uniq<PartialBlockForCheckpoint>(column_data, segment, state, block_manager);
 }
 
 void PartialBlockManager::RegisterPartialBlock(PartialBlockAllocation allocation) {
