+++ src/storage/table/CMakeLists.txt
@@ -8,6 +8,7 @@ add_library_unity(
   column_segment.cpp
   array_column_data.cpp
   list_column_data.cpp
+  in_memory_checkpoint.cpp
   update_segment.cpp
   persistent_table_data.cpp
   row_group.cpp
